+++ components/users/team.ts
@@ -16,8 +16,7 @@ export default class Team extends Component {
                 if (this.checkUser(userInfo.id, teamMember)) {
                     throw new Error(`User ${username} already on this team`);
                 }
-
-                const { response } = await this.helper.toPromise(this.api.teamApi, this.api.teamApi.teamsTeamIdUsersUserIdPost, teamInfo.data.id, userInfo.id,  { roleId: role } );
+                const { response } = await this.helper.toPromise(this.api.teamApi, this.api.teamApi.teamsTeamIdUsersUserIdPost, teamInfo.id, userInfo.id,  { roleId: role } );
                 if (!response.body) { 
                     throw new Error("Error adding user to team: invalid roleId");
                 }
@@ -31,7 +30,7 @@ export default class Team extends Component {
         }
     }
     
-    public async removeMember(username : string) {
+    public async removeMember(username: string) {
         const answer = await this.helper.inquirerPrompt([
             {
                 type: "confirm",
@@ -52,7 +51,7 @@ export default class Team extends Component {
                     throw new Error(`User ${username} not a member of this team`);
                 }
 
-                const { response } = await this.helper.toPromise(this.api.teamApi, this.api.teamApi.teamsTeamIdUsersUserIdDelete, teamInfo.data.id, userInfo.id);
+                const { response } = await this.helper.toPromise(this.api.teamApi, this.api.teamApi.teamsTeamIdUsersUserIdDelete, teamInfo.id, userInfo.id);
 
                 console.log(`Success remove ${username} from ${currentLogin}`);
             } else {
@@ -64,28 +63,54 @@ export default class Team extends Component {
         }
     }
 
-    private async getInfo(username : string) {
+    private async getInfo(username: string) {
         const currentLogin = this.helper.getProp("current_login") as string;
         const currentUserType = this.helper.getProp("current_user_type") as string;
 
         if (currentUserType !== "team") {
             throw new Error("Must be on team to do this operation");
         }
+
+        const requestTeamData =
+            await this.helper.toPromise(this.api.userApi, this.api.userApi.usersUserIdGet, currentLogin);
+
+        let teamInfo: any;
+        if (requestTeamData.response && requestTeamData.response.body) {
+            teamInfo = requestTeamData.response.body;
+        } else {
+            throw new Error("Cannot add user to team");
+        }
         
-        const team = await this.helper.toPromise(this.api.userApi, this.api.userApi.usersUserIdGet, currentLogin);
-        const { data } = await this.helper.toPromise(this.api.userApi, this.api.userApi.usersUserIdGet, username);
-        const member = await this.helper.toPromise(this.api.teamApi, this.api.teamApi.teamsTeamIdUsersGet, team.data.id);
+        const requestUserData =
+            await this.helper.toPromise(this.api.userApi, this.api.userApi.usersUserIdGet, username);
+
+        let userInfo: any;
+        if (requestUserData && requestUserData.response) {
+            userInfo = requestUserData.response.body;
+        } else {
+            throw new Error("Cannot add user to team");
+        }
+
+        const requestTeamMember =
+            await this.helper.toPromise(this.api.teamApi, this.api.teamApi.teamsTeamIdUsersGet, teamInfo.id);
+
+        let teamMember: any[];
+        if (requestTeamMember && requestTeamMember.response) {
+            teamMember = requestTeamMember.response.body;
+        } else {
+            throw new Error("Cannot add user to team");
+        }
 
         return {
-            teamInfo: team,
-            userInfo: data,
-            teamMember: member,
+            teamInfo,
+            userInfo,
+            teamMember,
             currentLogin
         };
     }
 
-    private checkUser(userId : string, member : any) : boolean {
-        const teamMember = member.response.body.map((x : JsonObject) => x.userId);
+    private checkUser(userId: string, member: any[]): boolean {
+        const teamMember = member.map((x: JsonObject) => x.userId);
         if (teamMember.indexOf(userId) > -1) {
             return true;
         }